API Flask has been installed and configured, to later transform it to Django using a virtual environment where it is executed in order to capture the versions and corroborate that it always runs well with the appropriate versions of everything installed.
Now there is a button in Quirk that what it does is to pass the
current url of that moment to Django. In Django the url that Quirk has
passed with its corresponding id is saved in the database. It sends a
response to Quirk with the id of the url and this one passes it to the
React application. The React application receives the id and sends a get
 with that id to the Django database to receive the associated url and
from here to be able to work with it.